ajourter column dans join table

+++ Server/monprojetcompanion/src/main/java/com/viseo/companion/evenements/dao/EventDAO.java
@@ -8,7 +8,6 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.ParameterExpression;
 import javax.persistence.criteria.Root;
 
 import org.springframework.stereotype.Repository;
@@ -23,11 +22,8 @@ public class EventDAO {
 	@PersistenceContext
 	EntityManager em;
 
-
-
-
 	@Transactional
-	public Event getEvent(long id){
+	public Event getEvent(int id){
 		return em.find(Event.class, id);
 	}
